set up a configuration file for run number and added function comments

+++ include/eventStruct.hh
@@ -7,7 +7,7 @@
 struct Event {
     uint32_t eventID;
     uint64_t timestamp;
-    std::vector<uint16_t> data; // Raw TDC hits, for example
+    std::vector<uint32_t> data; // Raw TDC hits, for example
 };
 
 #endif // EVENT_HH
